new file:   README.md 	new file:   README_EN.md 	new file:   benchmark/CMakeLists.txt 	new file:   benchmark/README.md 	new file:   benchmark/common/cmdline.hpp 	new file:   benchmark/common/timer.cc 	new file:   benchmark/common/timer.hpp 	new file:   benchmark/models/googlenet_benchmark.tmfile 	new file:   benchmark/models/inception_v3_benchmark.tmfile 	new file:   benchmark/models/mobilefacenets_benchmark.tmfile 	new file:   benchmark/models/mobilenet_benchmark.tmfile 	new file:   benchmark/models/mobilenet_v2_benchmark.tmfile 	new file:   benchmark/models/mobilenet_v3_benchmark.tmfile 	new file:   benchmark/models/mssd_benchmark.tmfile 	new file:   benchmark/models/resnet18_benchmark.tmfile 	new file:   benchmark/models/resnet50_benchmark.tmfile 	new file:   benchmark/models/retinaface_benchmark.tmfile 	new file:   benchmark/models/shufflenet_v2_benchmark.tmfile 	new file:   benchmark/models/squeezenet_v1.1_benchmark.tmfile 	new file:   benchmark/models/vgg16_benchmark.tmfile 	new file:   benchmark/models/yolov3_tiny_benchmark.tmfile 	new file:   benchmark/tm_benchmark.cc 	new file:   build/CMakeCache.txt 	new file:   build/CMakeFiles/3.22.1/CMakeASMCompiler.cmake 	new file:   build/CMakeFiles/3.22.1/CMakeCCompiler.cmake 	new file:   build/CMakeFiles/3.22.1/CMakeCXXCompiler.cmake 	new file:   build/CMakeFiles/3.22.1/CMakeDetermineCompilerABI_C.bin 	new file:   build/CMakeFiles/3.22.1/CMakeDetermineCompilerABI_CXX.bin 	new file:   build/CMakeFiles/3.22.1/CMakeSystem.cmake

+++ source/operator/prototype/slice_param.h
@@ -36,10 +36,10 @@ typedef struct slice_param
    // int axis;
    // int step;
    
-    int64_t begin;
-    int64_t end;
-    int64_t axis;
-    int64_t step;
+    int32_t begin;
+    int32_t end;
+    int32_t axis;
+    int32_t step;
     
     int ismxnet;
     int iscaffe;
